requirements.txt
glcontext==2.3.3
git+git://github.com/moderngl/moderngl@67c9d33#egg=moderngl
msgpack==1.0.2
multipledispatch==0.6.0
numpy==1.20.2
Pillow==8.2.0
pyrr==0.10.3
six==1.15.0
Wand==0.6.6
